requirements.txt
python-telegram-bot
selenium
webdriver-manager
selenium-stealth
beautifulsoup4